return port as int

+++ torrent/session.go
@@ -65,7 +65,7 @@ type Session struct {
 	torrentsByInfoHash map[dht.InfoHash][]*Torrent
 
 	mPorts         sync.RWMutex
-	availablePorts map[uint16]struct{}
+	availablePorts map[int]struct{}
 
 	mBlocklist         sync.RWMutex
 	blocklist          *blocklist.Blocklist
@@ -136,9 +136,9 @@ func New(cfg Config) (*Session, error) {
 			return nil, err
 		}
 	}
-	ports := make(map[uint16]struct{})
+	ports := make(map[int]struct{})
 	for p := cfg.PortBegin; p < cfg.PortEnd; p++ {
-		ports[p] = struct{}{}
+		ports[int(p)] = struct{}{}
 	}
 	bl := blocklist.New()
 	c := &Session{
@@ -367,9 +367,9 @@ func (s *Session) loadExistingTorrents(ids []string) {
 		t.webseedClient = &s.webseedClient
 		t.webseedSources = webseedsource.NewList(spec.URLList)
 		go s.checkTorrent(t)
-		delete(s.availablePorts, uint16(spec.Port))
+		delete(s.availablePorts, spec.Port)
 
-		t2 := s.newTorrent(t, id, uint16(spec.Port), spec.AddedAt, ann)
+		t2 := s.newTorrent(t, id, spec.Port, spec.AddedAt, ann)
 		s.log.Debugf("loaded existing torrent: #%d %s", id, t.Name())
 		loaded++
 		if hasStarted {
@@ -450,7 +450,7 @@ func (s *Session) AddTorrent(r io.Reader) (*Torrent, error) {
 	}
 	defer func() {
 		if err != nil {
-			s.releasePort(uint16(opt.Port))
+			s.releasePort(opt.Port)
 		}
 	}()
 	opt.Name = mi.Info.Name
@@ -490,7 +490,7 @@ func (s *Session) AddTorrent(r io.Reader) (*Torrent, error) {
 	if err != nil {
 		return nil, err
 	}
-	t2 := s.newTorrent(t, id, uint16(opt.Port), rspec.AddedAt, ann)
+	t2 := s.newTorrent(t, id, opt.Port, rspec.AddedAt, ann)
 	return t2, t2.Start()
 }
 
@@ -533,7 +533,7 @@ func (s *Session) addMagnet(link string) (*Torrent, error) {
 	}
 	defer func() {
 		if err != nil {
-			s.releasePort(uint16(opt.Port))
+			s.releasePort(opt.Port)
 		}
 	}()
 	opt.Name = ma.Name
@@ -565,7 +565,7 @@ func (s *Session) addMagnet(link string) (*Torrent, error) {
 	if err != nil {
 		return nil, err
 	}
-	t2 := s.newTorrent(t, id, uint16(opt.Port), rspec.AddedAt, ann)
+	t2 := s.newTorrent(t, id, opt.Port, rspec.AddedAt, ann)
 	return t2, t2.Start()
 }
 
@@ -595,7 +595,7 @@ func (s *Session) add() (*options, *filestorage.FileStorage, string, error) {
 	}
 	return &options{
 		id:         id,
-		Port:       int(port),
+		Port:       port,
 		Resumer:    res,
 		Blocklist:  s.blocklist,
 		PieceCache: s.pieceCache,
@@ -604,7 +604,7 @@ func (s *Session) add() (*options, *filestorage.FileStorage, string, error) {
 	}, sto, id, nil
 }
 
-func (s *Session) newTorrent(t *torrent, id string, port uint16, addedAt time.Time, ann *dhtAnnouncer) *Torrent {
+func (s *Session) newTorrent(t *torrent, id string, port int, addedAt time.Time, ann *dhtAnnouncer) *Torrent {
 	t2 := &Torrent{
 		session:      s,
 		torrent:      t,
@@ -622,7 +622,7 @@ func (s *Session) newTorrent(t *torrent, id string, port uint16, addedAt time.Ti
 	return t2
 }
 
-func (s *Session) getPort() (uint16, error) {
+func (s *Session) getPort() (int, error) {
 	s.mPorts.Lock()
 	defer s.mPorts.Unlock()
 	for p := range s.availablePorts {
@@ -632,7 +632,7 @@ func (s *Session) getPort() (uint16, error) {
 	return 0, errors.New("no free port")
 }
 
-func (s *Session) releasePort(port uint16) {
+func (s *Session) releasePort(port int) {
 	s.mPorts.Lock()
 	defer s.mPorts.Unlock()
 	s.availablePorts[port] = struct{}{}
